sources: Add optional `user` parameter to media sources. (#133)
Sources can now receive a `context` parameter on the Search and Get
actions, in addition to Import actions. Sources can opt in by setting
their `api` property to 2, so sources that don't need it will still work.

The `context` parameter is an object that currently has `uw`, `source`
and `user` properties.