Move waitlist into core (#250)
I tried this a bunch of times, integrating the waitlist nicely with the
existing booth code, but always got stuck. So instead, this is a simple
copy-paste job from http-api, which is at least a start. This makes it
easier to apply small changes in the future to work towards a nicer
integration.