playlists: use a single aggregation query to fetch playlist items
Previously, pagination was done in JavaScript and we needed separate
methods to do filtering.

MongoDB has a `$lookup` aggregation stage to join data from a different
collection. We can use it to associate playlist items with the playlist
on the DB side, and later to get the media object for each playlist
item.

MongoDB has a `$facet` aggregation stage that allows you to split up the
stages: one to only count _all_ filter matches, and one to return the
playlist items that match the filter and survived pagination.