Improve `getPlaylistsContainingAnyMedia` performance
I thought this `$lookup.pipeline` solution worked pretty well, but I
think it doesn't let MongoDB use the indices on the `playlistitems`
collection. Also in the past I did `$project` early because I thought it
would be faster to have less data for the remaining steps of the
pipeline, but it's actually very slow when you still have lots of
documents in it.

So, now this uses the simple `$lookup` syntax that only has to look at
an indexed property in `playlistitems`, and moves `$project` to the end.

This cuts it down to a query that's maybe 200ms on wlk.yt (with a few
hundred thousand playlist items).